+++ pyfixest/did/__init__.py
@@ -1,0 +1,7 @@
+print(
+    """
+      You have loaded the 'pyfixest.did' module. While every function is tested in `tests/test_did.py`,
+      the module is not yet as thoroughly tested as I would like. So please use it with caution and
+      provide feedback in case you stumble over any bugs!
+      """
+)